% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allupstream.R
\name{allupstream}
\alias{allupstream}
\title{Return a vector of all subcatchments upstream of a specified catchment}
\usage{
allupstream(cat.hier, catchname)
}
\arguments{
\item{cat.hier}{a dataframe containing catchment id and next downstream (nextds) id fields}

\item{catchname}{a dataframe containing catchment id for which a vector of upstream catchment
ids will be returned.}
}
\value{
a vector of downstream catchment ids
}
\description{
Return a vector of all subcatchments upstream of a specified catchment
}
\note{
Function depends on the next downstream field in a stream network 'hierarchy' table (dataframe).
When used in conjunction with list_all_upstream it is possible to get a list of catchments upstream
of a vector of sites. This can then be used to support further aggregation of environmental variables
for contributing catchments upstream of a list of catchments of interest (for example for aggregating rainfall).
}
\examples{

}

