% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_catch.R
\name{build_catch}
\alias{build_catch}
\title{function for building a data frame for each site in sitelist}
\usage{
build_catch(sitelist, hierarchy, catchstats)
}
\arguments{
\item{sitelist}{a single row from a dataframe with list of sites with corresponding d2ol values}

\item{hierarchy}{must have fields site, nextds and d2ol (can be < sitelist$d2ol)
containing all the data in catchstats for all upstream catchments}

\item{catchstats}{?}
}
\value{
?
}
\description{
function for building a data frame for each site in sitelist
}
\note{
?
}
\examples{
#data(mwcats)

#to be added.

}

