% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_catch_pxls.R
\name{build_catch_pxls}
\alias{build_catch_pxls}
\title{function for building a data frame for a single site in sitelist}
\usage{
build_catch_pxls(sitelist, hierarchy)
}
\arguments{
\item{sitelist}{a single row from a dataframe with list of sites with corresponding d2ol values}

\item{hierarchy}{must have fields site, nextds and d2ol (can be < sitelist$d2ol)
containing all the data in catchstats for all upstream catchments}
}
\value{
a dataframe
}
\description{
function for building a data frame for a single site in sitelist
}
\note{
using hierarchy and the allupstream function.
it assembles a dataframe from the vectors subcID, luno, d2ol, d2str, and fa
}
\examples{
data(mwcats)

#to be added.

}

