% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_stats.R
\name{catch_stats}
\alias{catch_stats}
\title{Calculate full catchment statistics by summing field(s) value(s) for all upstream catchments}
\usage{
catch_stats(sitelist, hierarchy, subcatchment.data, FUN, ...)
}
\arguments{
\item{sitelist}{a vector of sites for which compiled statistics are to be calculated.}

\item{hierarchy}{a dataframe containing catchment id and next downstream (nextds) id fields.}

\item{subcatchment.data}{a vector or dataframe containing variables for which statistics are sought.}

\item{FUN}{an unquoted scalar function used to summarise each field (e.g. min, mean, max).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector or dataframe of aggregate values
}
\description{
Calculate full catchment statistics by summing field(s) value(s) for all upstream catchments
}
\note{
function to calculate full catchment statistics by summing each field
of the table (or vector) subcatchment.data (length = no. of rows in hierarchy)
For each site in sitelist, statistics as per scalar function FUN
for all upstream catchments are deriveded using hierarchy in the
function allupstream.
}
\examples{
data(mwcats)

catch_stats(sitelist=mwcats$site[1:5], hierarchy = mwcats[,c(1:2)],
subcatchment.data=mwcats$scstrlen, FUN=sum)

}
