% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_awap.R
\name{download_awap}
\alias{download_awap}
\title{Code to download monthly AWAP grids to a specified directory}
\usage{
download_awap(loc_dir = NULL)
}
\arguments{
\item{loc_dir}{The local directory to which the files will be downloaded and unzipped}
}
\value{
Downloads CSIRO AWAP files to the specified directory ready for further use
}
\description{
Code to download monthly AWAP grids to a specified directory
}
\note{
The files will be downloaded to the directory specified (defaults to the
current working directory) and unzipped. The function first checks to see which zip
files have previously been downloaded, and for this reason, if deleting files to save
disk space it is best to delete the unzipped files in the directory '/AWAP/Run26j/FWDis/'.
}
\examples{
download_awap()
}

