% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findnds.R
\name{findnds}
\alias{findnds}
\title{Identify the next downstream site from a set of candidates}
\usage{
findnds(cat.hier, catchnames, candidates)
}
\arguments{
\item{cat.hier}{a dataframe containing catchment id and next downstream (nextds) id fields}

\item{catchnames}{a vector of catchment ids for which a a list of downstream catchments
ids will be returned.}
}
\value{
a list of downstream catchment ids for each catchment in catchnames
}
\description{
Identify the next downstream site from a set of candidates
}
\note{
Function depends on the next downstream field in a stream network 'hierarchy' table (dataframe).
Can be used to support further aggregation of environmental variables for sub-catchments downstream
of a list of catchments of interest (e.g. for calculating barrier numbers).
}
\examples{

}

