% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_nds.R
\name{list_all_nds}
\alias{list_all_nds}
\title{Identify the next downstream site for each site in a list of sites.}
\usage{
list_all_nds(cat.hier, catchnames, candidates)
}
\arguments{
\item{cat.hier}{a dataframe containing catchment id and next downstream (nextds) id fields}

\item{catchnames}{a vector of catchment ids (sites) for which a a list of downstream catchments
ids will be returned.}

\item{candidates}{a vector of candidate catchment ids. Note this need not be a complete hierarchy}
}
\value{
a list of downstream catchment ids for each catchment in catchnames. Returns -1 where there
is no downstream site
}
\description{
Identify the next downstream site for each site in a list of sites.
}
\note{
Function depends on the next downstream field in a stream network 'hierarchy' table (dataframe).
Can be used to support further aggregation of environmental variables for sub-catchments downstream
of a list of catchments of interest (e.g. for calculating barrier numbers).
}
\examples{
data(mwcats)

#find next downstream sites from a list of candidates for a list of sites
data(mwcats)

list_all_nds(cat.hier = mwcats[,c(1:2)], catchnames = mwcats$site[1:5],
candidates=sample(mwcats$site, size = 4000))

}

