% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkStats.R
\name{networkstats}
\alias{networkstats}
\title{summarise river network character between site pairs}
\usage{
networkstats(site1, site2, cat_data, catid_col, nextds_col, env_vars, FUN,
  ...)
}
\arguments{
\item{site1}{a vector of sites for which all pairs are sought}

\item{site2}{a second vector of sites for which all pairs are sought}

\item{cat_data}{the name of the dataframe containing the catchment data}

\item{catid_col}{the column containing the catchment ids}

\item{nextds_col}{the column containing the next downstream catchment ids}

\item{env_vars}{quoted vector of environmental variables for which summary stats are required}

\item{FUN}{quoted scalar function to apply to env.vars (e.g. 'sum')}

\item{...}{additional arguments to FUN}
}
\value{
a dataframe with site pairs and network characteristics
}
\description{
summarise river network character between site pairs
}
\note{
...
}
\examples{
data(mwcats)

all_pairs(mwcats$site[1:5])

}
