% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairReaches.R
\name{pairReaches}
\alias{pairReaches}
\title{Create a list of all reaches between two sites in a river network (inclusive of those sites)}
\usage{
pairReaches(hierarchy, site1, site2)
}
\arguments{
\item{hierarchy}{a dataframe containing catchment id and next downstream (nextds) id fields.}

\item{site1}{a site in a river network}

\item{site2}{a site in the same river network as site1 (can be up or downstream)}
}
\description{
Create a list of all reaches between two sites in a river network (inclusive of those sites)
}
\note{
Produces a vector of sites between the site pair
}
\examples{
data(mwcats)

#create a dataframe of pair reaches
pairReaches(hierarchy=mwcats[,c(1:2)], mwcats$site[1], mwcats$site[5])

#create a dataframe with all pairs from a vector of sites
df<- all_pairs(mwcats$site[1:5])

#create a list of sites between each site pair
apply(df,1 , function(x) pairReaches(hierarchy=mwcats[,c(1:2)], x[[1]],x[[2]]))

}
