% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverreach.R
\name{riverreach}
\alias{riverreach}
\title{returns list of subcatchments between 2 subcatchments on the same river ordered upstream to downstream}
\usage{
riverreach(hierarchy, upstream, downstream, riverdist = NULL)
}
\arguments{
\item{hierarchy}{a dataframe containing catchment id and next downstream (nextds) id fields.
Hierarchy must have fields 'site' and 'nextds'}

\item{upstream}{the upstream subcatchment id}

\item{downstream}{the downstream subcatchment id}

\item{riverdist}{?}
}
\value{
Returns list of subcatchments between 2 subcatchments
on the same river.
}
\description{
returns list of subcatchments between 2 subcatchments on the same river ordered upstream to downstream
}
\note{
riverdist, if used, should be the column number of the riverdistance
measure in hierarchy.
Returns the following error if downstream is not downstream of upstream:
Error in while (y != downstream) { : argument is of length zero' }
Note that unlike pairReaches riverreach preserves ordering but cannot trace down and then up tributaries.
}
\examples{
data(mwcats)
#function needs checking.
#riverreach(hierarchy = mwcats[,c(1:2)], upstream='YARR2415', downstream='YARR001')

}
