% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findnds.R
\name{findnds}
\alias{findnds}
\title{Identify the next downstream site from a set of candidates}
\usage{
findnds(hierarchy, catchname, candidates)
}
\arguments{
\item{hierarchy}{a dataframe containing catchment id and next downstream (nextds) id fields}

\item{catchname}{a catchment id for which a a list of downstream catchments
ids will be returned.}

\item{candidates}{a vector of candidate catchment ids. Note this need not be a complete hierarchy}
}
\value{
a list of downstream catchment ids for each catchment in catchnames. Returns -1 where there
is no downstream site
}
\description{
Identify the next downstream site from a set of candidates
}
\note{
Function depends on the next downstream field in a stream network 'hierarchy' table (dataframe).
Can be used to support further aggregation of environmental variables for sub-catchments downstream
of a list of catchments of interest (e.g. for calculating barrier numbers).
}
\examples{
data(mwcats)

#find site next downstream of the first site in the catchment list

findnds(hierarchy = mwcats[,c(1:2)], catchname = mwcats$site[1], candidates=mwcats$site)

}
