% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_downstream.R
\name{list_all_downstream}
\alias{list_all_downstream}
\title{Identify all the subcatchments downstream of a list of specified catchments}
\usage{
list_all_downstream(cat.hier, catchnames)
}
\arguments{
\item{cat.hier}{a dataframe containing catchment id and next downstream (nextds) id fields}

\item{catchnames}{a vector of catchment ids for which a vector of next downstream catchment
ids will be returned.}
}
\value{
a vector of next downstream catchment ids for each catchment in catchnames
}
\description{
Identify all the subcatchments downstream of a list of specified catchments
}
\note{
Function depends on the next downstream field in a stream network 'hierarchy' table (dataframe).
}
\examples{

}

