% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkStats.R
\name{networkstats}
\alias{networkstats}
\title{summarise river network character between site pairs}
\usage{
networkstats(site1, site2, FUN = "max", env.vars = NULL)
}
\arguments{
\item{site1}{a vector of sites for which all pairs are sought}

\item{site2}{a second vector of sites for which all pairs are sought}

\item{FUN}{quoted scalar function to apply to env.vars (e.g. 'sum')}

\item{env.vars}{environmental variables for which summary stats are required}
}
\value{
a dataframe with site pairs and network characteristics
}
\description{
summarise river network character between site pairs
}
\note{
...
}
\examples{
#data(mwcats)

#all_pairs(mwcats$site[1:5])

}

